+++ src/rez/bind/hello_world.py
@@ -12,7 +12,7 @@ case, and for that reason we do not want any dependencies.
 from __future__ import absolute_import, print_function
 
 from rez.package_maker import make_package
-from rez.vendor.version.version import Version
+from rez.version import Version
 from rez.utils.lint_helper import env
 from rez.utils.execution import create_executable_script, ExecutableScriptMode
 from rez.bind._utils import make_dirs, check_version
