+++ src/rez/tests/test_copy_package.py
@@ -16,7 +16,7 @@ from rez.build_system import create_build_system
 from rez.resolved_context import ResolvedContext
 from rez.packages import get_latest_package
 from rez.package_copy import copy_package
-from rez.vendor.version.version import VersionRange
+from rez.version import VersionRange
 from rez.tests.util import TestBase, TempdirMixin
