+++ src/rez/tests/test_packages.py
@@ -20,8 +20,8 @@ from rez.tests.util import TestBase, TempdirMixin
 from rez.utils.formatting import PackageRequest
 from rez.utils.sourcecode import SourceCode
 import unittest
-from rez.vendor.version.version import Version
-from rez.vendor.version.util import VersionError
+from rez.version import Version
+from rez.version import VersionError
 from rez.utils.filesystem import canonical_path
 import shutil
 import os.path
