+++ src/rezgui/widgets/VariantCellWidget.py
@@ -9,8 +9,8 @@ from rezgui.mixins.ContextViewMixin import ContextViewMixin
 from rez.packages import PackageSearchPath
 from rez.package_filter import PackageFilterList
 from rez.resolved_context import PatchLock, get_lock_request
-from rez.vendor.version.requirement import RequirementList
-from rez.vendor.version.version import VersionRange
+from rez.version import RequirementList
+from rez.version import VersionRange
 from functools import partial
