+++ src/rezgui/widgets/VariantHelpWidget.py
@@ -6,7 +6,7 @@ from Qt import QtCore, QtWidgets
 from rezgui.util import create_pane, get_icon_widget
 from rezgui.mixins.ContextViewMixin import ContextViewMixin
 from rezgui.widgets.PackageLoadingWidget import PackageLoadingWidget
-from rez.vendor.version.version import VersionRange
+from rez.version import VersionRange
 from rez.package_help import PackageHelp
 from functools import partial
