+++ src/rezgui/widgets/VariantVersionsTable.py
@@ -7,7 +7,7 @@ from rezgui.mixins.ContextViewMixin import ContextViewMixin
 from rez.package_filter import PackageFilterList
 from rezgui.util import get_timestamp_str, update_font, get_icon_widget, create_pane
 from rez.packages import iter_packages
-from rez.vendor.version.version import VersionRange
+from rez.version import VersionRange
 
 
 class VariantVersionsTable(QtWidgets.QTableWidget, ContextViewMixin):
